requirements.txt
altair==5.5.0
df-explorer==0.0.3
extra-streamlit-components==0.1.71
importlib-metadata==8.6.1
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
numerize==0.12
numpy==2.2.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pipreqs==0.5.0
plotly==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.15.1
seaborn==0.13.2
streamlit-card==1.0.2
streamlit-extras==0.5.0
streamlit-keyup==0.3.0
streamlit-option-menu==0.4.0
toml==0.10.2
